feat: redesign homepage — merge about page content, fix title, remove redirect on history/fav tap
